Extract stake actions creation into separate file
