+++ packages/app/src/features/actions/flavours/stake/types.ts
@@ -4,16 +4,16 @@ import { Token } from '@/domain/types/Token'
 
 export interface StakeObjective {
   type: 'stake'
-  token: Token // any supported input token
-  amount: NormalizedUnitNumber // amount of input token, not necessarily stake amount (e.g. when input is sdai/susds)
+  token: Token // any supported input token (e.g. dai, usds, usdc, sdai, susds, ...)
+  amount: NormalizedUnitNumber // amount of input token, not necessarily stake amount (in case when input is savings token)
   farm: CheckedAddress
   isMax: boolean
 }
 
 export interface StakeAction {
   type: 'stake'
-  stakingToken: Token // usds or other supported staking token
-  stakeAmount: NormalizedUnitNumber // amount of staking token (usds or other supported staking token)
+  stakingToken: Token // e.g. usds for Sky farms
+  stakeAmount: NormalizedUnitNumber
   rewardToken: Token
   farm: CheckedAddress
 }
